add custom usePersistedReducer
